moved resolver code to separate files
