+++ application/gqlgen/organization.go
@@ -1,0 +1,34 @@
+package gqlgen
+
+import (
+	"context"
+	"github.com/silinternational/handcarry-api/domain"
+	"github.com/silinternational/handcarry-api/models"
+)
+
+func (r *Resolver) Organization() OrganizationResolver {
+	return &organizationResolver{r}
+}
+
+type organizationResolver struct{ *Resolver }
+
+func (r *organizationResolver) URL(ctx context.Context, obj *models.Organization) (*string, error) {
+	if obj == nil {
+		return nil, nil
+	}
+	return GetStringFromNullsString(obj.Url), nil
+}
+
+func (r *organizationResolver) CreatedAt(ctx context.Context, obj *models.Organization) (*string, error) {
+	if obj == nil {
+		return nil, nil
+	}
+	return domain.ConvertTimeToStringPtr(obj.CreatedAt), nil
+}
+
+func (r *organizationResolver) UpdatedAt(ctx context.Context, obj *models.Organization) (*string, error) {
+	if obj == nil {
+		return nil, nil
+	}
+	return domain.ConvertTimeToStringPtr(obj.UpdatedAt), nil
+}
