+++ application/gqlgen/post.go
@@ -1,10 +1,147 @@
 package gqlgen
 
 import (
+	"context"
 	"fmt"
+	"github.com/99designs/gqlgen/graphql"
 	"github.com/silinternational/handcarry-api/domain"
 	"github.com/silinternational/handcarry-api/models"
+	"github.com/vektah/gqlparser/gqlerror"
 )
+
+func (r *Resolver) Post() PostResolver {
+	return &postResolver{r}
+}
+
+type postResolver struct{ *Resolver }
+
+func (r *postResolver) Type(ctx context.Context, obj *models.Post) (PostType, error) {
+	if obj == nil {
+		return "", nil
+	}
+	return PostType(obj.Type), nil
+}
+
+func (r *postResolver) Organization(ctx context.Context, obj *models.Post) (*models.Organization, error) {
+	if obj == nil {
+		return nil, nil
+	}
+	organization, err := obj.GetOrganization(GetRequestFields(ctx))
+	if err != nil {
+		return nil, fmt.Errorf("error retrieving Organization data for post %v: %v", obj.ID, err)
+	}
+	return &organization, nil
+}
+
+func (r *postResolver) Description(ctx context.Context, obj *models.Post) (*string, error) {
+	if obj == nil {
+		return nil, nil
+	}
+	return GetStringFromNullsString(obj.Description), nil
+}
+
+func (r *postResolver) Destination(ctx context.Context, obj *models.Post) (*string, error) {
+	if obj == nil {
+		return nil, nil
+	}
+	return GetStringFromNullsString(obj.Destination), nil
+}
+
+func (r *postResolver) Origin(ctx context.Context, obj *models.Post) (*string, error) {
+	if obj == nil {
+		return nil, nil
+	}
+	return GetStringFromNullsString(obj.Origin), nil
+}
+
+func (r *postResolver) NeededAfter(ctx context.Context, obj *models.Post) (*string, error) {
+	if obj == nil {
+		return nil, nil
+	}
+	return domain.ConvertTimeToStringPtr(obj.NeededAfter), nil
+}
+
+func (r *postResolver) NeededBefore(ctx context.Context, obj *models.Post) (*string, error) {
+	if obj == nil {
+		return nil, nil
+	}
+	return domain.ConvertTimeToStringPtr(obj.NeededBefore), nil
+}
+
+func (r *postResolver) Threads(ctx context.Context, obj *models.Post) ([]*models.Thread, error) {
+	return nil, nil
+}
+
+func (r *postResolver) CreatedAt(ctx context.Context, obj *models.Post) (*string, error) {
+	if obj == nil {
+		return nil, nil
+	}
+	return domain.ConvertTimeToStringPtr(obj.CreatedAt), nil
+}
+
+func (r *postResolver) UpdatedAt(ctx context.Context, obj *models.Post) (*string, error) {
+	if obj == nil {
+		return nil, nil
+	}
+	return domain.ConvertTimeToStringPtr(obj.UpdatedAt), nil
+}
+
+func (r *postResolver) MyThreadID(ctx context.Context, obj *models.Post) (*string, error) {
+	if obj == nil {
+		return nil, nil
+	}
+
+	currentUser := models.GetCurrentUserFromGqlContext(ctx, TestUser)
+
+	thread, err := models.FindThreadByPostIDAndUserID(obj.ID, currentUser.ID)
+	if err != nil {
+		return nil, err
+	}
+
+	threadUuid := thread.Uuid.String()
+	if threadUuid == domain.EmptyUUID {
+		return nil, nil
+	}
+
+	return &threadUuid, nil
+}
+
+func (r *queryResolver) Posts(ctx context.Context) ([]*models.Post, error) {
+
+	db := models.DB
+	var posts []*models.Post
+	cUser := models.GetCurrentUserFromGqlContext(ctx, TestUser)
+	if err := db.Where("organization_id IN (?)", cUser.GetOrgIDs()...).All(&posts); err != nil {
+		graphql.AddError(ctx, gqlerror.Errorf("Error getting posts: %v", err.Error()))
+		return []*models.Post{}, err
+	}
+
+	for _, p := range posts {
+		e := p.QueryRelatedUsers(GetSelectFieldsFromRequestFields(UserSimpleFields(), GetRequestFields(ctx)))
+		if e != nil {
+			return posts, e
+		}
+	}
+
+	return posts, nil
+}
+
+func (r *queryResolver) Post(ctx context.Context, id *string) (*models.Post, error) {
+	post := models.Post{}
+	cUser := models.GetCurrentUserFromGqlContext(ctx, TestUser)
+
+	if err := models.DB.Where("organization_id IN (?)", cUser.GetOrgIDs()...).Where("uuid = ?", id).First(&post); err != nil {
+		graphql.AddError(ctx, gqlerror.Errorf("Error getting post: %v", err.Error()))
+		return &models.Post{}, err
+	}
+
+	e := post.QueryRelatedUsers(GetSelectFieldsFromRequestFields(UserSimpleFields(), GetRequestFields(ctx)))
+	if e != nil {
+		return &post, e
+	}
+
+	return &post, nil
+}
 
 // ConvertGqlNewPostToDBPost does what its name says, but also ...
 func ConvertGqlNewPostToDBPost(gqlPost NewPost, createdByUser models.User) (models.Post, error) {
